Move tests under row_generation
