+++ test/row_generation/__init__.py
@@ -1,0 +1,1 @@
+from .GeneratorTestBase import GeneratorTestBase
